+++ lib/fancy_sticky_notes.dart
@@ -1,4 +1,5 @@
 export 'src/sticky_note_style.dart';
+export 'src/styled_text_controller.dart';
 export 'src/sticky_note.dart';
 export 'src/sticky_note_painter.dart';
 export 'src/sticky_note_board.dart';